WSTEAM1-705: Clean up
