+++ ws-nextjs-app/pages/[service]/live/[id]/LivePageLayout.tsx
@@ -11,10 +11,7 @@ import Link from 'next/link';
 import MetadataContainer from '../../../../../src/app/components/Metadata';
 import LinkedDataContainer from '../../../../../src/app/components/LinkedData';
 import Stream from './Stream';
-import {
-  StreamContext,
-  StreamProvider,
-} from './Stream/BackToLatest/stream-provider';
+import { StreamProvider } from './Stream/BackToLatest/stream-provider';
 import Header from './Header';
 import KeyPoints from './KeyPoints';
 
@@ -43,78 +40,27 @@ type ComponentProps = {
   post: string | null;
 };
 
-const MyLinks = () => {
-  // const linkOnClickOverride = event => {
-  //   const tldRegex = /(.co.uk|.com)/;
-  //   const targetUrl = new URL(event.target.href);
-  //   const currentUrl = new URL(window.location.href);
-
-  //   // TLDs and .app are removed when comparing to avoid redirect loop for users
-  //   const isOnSameHost =
-  //     targetUrl.host.replace(tldRegex, '') ===
-  //     currentUrl.host.replace(tldRegex, '');
-  //   const isOnSamePath =
-  //     targetUrl.pathname === currentUrl.pathname.replace('.app', '');
-
-  //   const postSearchParam = targetUrl.searchParams.get('post');
-  //   const newPath = `${currentUrl.pathname}?${targetUrl.searchParams}${targetUrl.hash}`;
-
-  //   // only override behaviour if on same page and has post
-  //   if (isOnSameHost && isOnSamePath && postSearchParam) {
-  //     event.preventDefault();
-  //     setPost(postSearchParam);
-  //     window.history.pushState(null, null, newPath);
-  //   }
-
-  //   const linkTrackingOverride = {
-  //     group: {
-  //       name: 'Summary Points'
-  //     }
-  //   };
-
-  //   const propObj = {
-  //     headingLevel: 2,
-  //     blocksOverride: data,
-  //     isContained: true,
-  //     isFullWidth: true,
-  //     headingFontScale: 'indexHeadlineMedium',
-  //     overrides: {
-  //       linkTrackingOverride,
-  //       linkOnClickOverride
-  //     }
-  //   };
-
-  //   return (
-  //     <>
-  //       <Heading fontScale="indexHeadlineMedium" level="2" type="section-heading">
-  //         {'Summary'}
-  //       </Heading>
-  //       <EditorialText {...propObj} />
-  //     </>
-  //   );
-  // };
-  // /**
-  //  * Summary Points module for showing Editorial Text retrieved from Tipo
-  //  * @type {typeof import('./summary-points').SummaryPoints}
-  //  */
-  // export const SummaryPoints = ({ data, setPost }) => {
-  //   return (
-  //     <SummaryPointsWrapper hasData={data?.length > 0} id="summaryPoints">
-  //       {Boolean(data?.length) && SummaryPointsContainer({ data, setPost })}
-  //     </SummaryPointsWrapper>
-  //   );
-  // };
-  // };
+const FakeKeyPointLinks = () => {
+  // We need to use the 'no scroll' link which ends in #post for accessibility
+  // However if we add an onClickOverride to the Summary Points then we could mimic the behaviour of the 'Scroll' link
+  // See the Dropbox Paper for PS Web implementation
 
   return (
     <>
-      {/* <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3A3b133574-88dc-41e0-9d90-0d3e847adba3#post"> */}
       <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3A3b133574-88dc-41e0-9d90-0d3e847adba3">
-        Link for post 39 (Page 1)
+        Link for post 39 (Page 1) - Scroll
+      </Link>
+      <br />
+      <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3A3b133574-88dc-41e0-9d90-0d3e847adba3#post">
+        Link for post 39 (Page 1) - No scroll
+      </Link>
+      <br />
+      <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3Aba735203-6eff-4768-83ce-74098a3ee92a">
+        Link for post 32 (Page 2) - Scroll
       </Link>
       <br />
       <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3Aba735203-6eff-4768-83ce-74098a3ee92a#post">
-        Link for post 32 (Page 2)
+        Link for post 32 (Page 2) - No scroll
       </Link>
       <br />
       <Link href="http://localhost:7081/pidgin/live/c07zr0zwjnnt?post=asset%3Ab14bc99c-eb76-47ef-a716-f4ce97ff1349#post">
@@ -125,12 +71,6 @@ const MyLinks = () => {
 };
 
 const LivePage = ({ pageData, post }: ComponentProps) => {
-  const streamRef = useRef(null);
-  const hiddenHeadlineRef = useRef(null);
-
-  console.log("I'm the streamRef", streamRef);
-  console.log("I'm the hiddenHeadlineRef", hiddenHeadlineRef);
-
   const { lang, translations } = useContext(ServiceContext);
   const {
     title,
@@ -141,6 +81,7 @@ const LivePage = ({ pageData, post }: ComponentProps) => {
     liveTextStream,
     atiAnalytics,
   } = pageData;
+  const streamRef = useRef(null);
 
   const { index: activePage, total: pageCount } =
     liveTextStream?.content?.data?.page || {};
@@ -201,21 +142,17 @@ const LivePage = ({ pageData, post }: ComponentProps) => {
             )}
           </div>
           <div css={styles.secondSection}>
-            <MyLinks />
+            <FakeKeyPointLinks />
             <StreamProvider
               streamRef={streamRef}
               post={post}
-              setPage={undefined}
-              setPost={undefined}
               activePage={activePage}
-              hiddenHeadlineRef={hiddenHeadlineRef}
             >
               <Stream
                 streamContent={liveTextStream.content}
                 contributors={liveTextStream.contributors}
                 post={post}
                 streamRef={streamRef}
-                hiddenHeadlineRef={hiddenHeadlineRef}
               />
             </StreamProvider>
           </div>
