+++ ws-nextjs-app/pages/[service]/live/[id]/Post/index.tsx
@@ -2,7 +2,6 @@
 import React, { useContext } from 'react';
 import { jsx } from '@emotion/react';
 import pathOr from 'ramda/src/pathOr';
-import Link from 'next/link';
 import { OptimoBlock } from '#models/types/optimo';
 import Heading from '#app/components/Heading';
 import Text from '#app/components/Text';
@@ -16,8 +15,6 @@ import { ServiceContext } from '#app/contexts/ServiceContext';
 import isTenHoursAgo from '#app/lib/utilities/isTenHoursAgo';
 import TimeStampContainer from '#app/legacy/psammead/psammead-timestamp-container/src';
 import SocialEmbedContainer from '#app/legacy/containers/SocialEmbed';
-import useLocation from '#hooks/useLocation';
-import makeRelativeUrlPath from '#lib/utilities/makeRelativeUrlPath';
 import styles from './styles';
 import {
   Post as PostType,
@@ -25,18 +22,8 @@ import {
   ComponentToRenderProps,
 } from './types';
 import BackToLatestPost from '../Stream/BackToLatest/back-to-latest-post';
-
-const CopyToClipboard = (toCopy: string) => {
-  const el = document.createElement(`textarea`);
-  el.value = toCopy;
-  el.setAttribute(`readonly`, ``);
-  el.style.position = `absolute`;
-  el.style.left = `-9999px`;
-  document.body.appendChild(el);
-  el.select();
-  document.execCommand(`copy`);
-  document.body.removeChild(el);
-};
+import getShareURL from './helpers/getShareUrl';
+import CopyToClipboard from './helpers/copyToClipboard';
 
 const PostBreakingNewsLabel = ({
   isBreakingNews,
@@ -194,19 +181,6 @@ const Post = ({ post }: { post: PostType }) => {
 
   const isBreakingNews = pathOr(false, ['options', 'isBreakingNews'], post);
   const timestamp = post?.dates?.curated ?? '';
-
-  // could this logic sit in the BFF and be passed through?
-  // is there a chance the POST asset ID will change? Ask the live team.
-  const getShareURL = ({ urn, scroll }: { urn: string; scroll: boolean }) => {
-    const myURL = `http://localhost:7081/pidgin/live/c07zr0zwjnnt`;
-    // replace this with URL builder like those in src/app/lib/utilities/getUrlHelpers
-    const cleanedURN = urn.split(':').pop();
-    const shareURL = scroll
-      ? `${myURL}?post=asset%3A${cleanedURN}`
-      : `${myURL}?post=asset%3A${cleanedURN}#post`;
-
-    return shareURL;
-  };
 
   return (
     <article css={styles.postContainer} id={postURN}>
