+++ ws-nextjs-app/pages/[service]/live/[id]/Stream/BackToLatest/stream-provider.tsx
@@ -1,44 +1,25 @@
 import React, { PropsWithChildren } from 'react';
 
 export type StreamContextProps = {
-  streamRef: any;
-  post: any;
-  setPage: any;
-  setPost: any;
+  streamRef: unknown; // HTML element/ null?
+  post: string | null; // e.g. asset:fd4643b5-191b-4794-a7ac-59b18c322c35
   activePage: number | undefined;
-  hiddenHeadlineRef: any;
 };
 
 export const StreamContext = React.createContext<StreamContextProps>(
   {} as StreamContextProps,
 );
 
-export type StreamProviderProps = {
-  streamRef: any; // HTML element?
-  post: string | null; // e.g. asset:fd4643b5-191b-4794-a7ac-59b18c322c35
-  setPage: any;
-  setPost: any;
-  children: any;
-  activePage: number | undefined;
-  hiddenHeadlineRef: any;
-};
-
 export const StreamProvider = ({
   streamRef,
   post,
-  setPage,
-  setPost,
   children,
   activePage,
-  hiddenHeadlineRef,
-}: PropsWithChildren<StreamProviderProps>) => {
+}: PropsWithChildren<StreamContextProps>) => {
   const value = {
     streamRef,
     post,
-    setPage,
-    setPost,
     activePage,
-    hiddenHeadlineRef,
   };
 
   return (
