--- ws-nextjs-app/pages/[service]/live/[id]/Stream/hooks/myhook.ts
@@ -1,14 +1,0 @@
-// import { useEffect } from 'react';
-
-// const MyHook = (postId: string | null) => {
-//   useEffect(() => {
-//     if (postId) {
-//       const element = document.getElementById(postId);
-
-//       element?.scrollIntoView();
-//       (element as HTMLElement)?.focus();
-//     }
-//   }, [postId]);
-// };
-
-// export default MyHook;
